Create Week 7 Graph and bfs and dfs
13. Write the pseudocode for an unweighted graph data structure. You either use an adjacency
matrix or an adjacency list approach. Also, write a function to add a new node and a
function to add an edge. Following that, implement the graph you have designed in the
programming language of your choice. You may use your own linked list from previous labs,
the STL LL, or use a simple fixed size array (10 elements would be fine).
14. Implement BFS and DFS traversals for the above graph. Save the nodes traversed in
sequence to a text file.